Fix build on Windows for x64.

+++ os/windows/AudioOutputWave.cpp
@@ -101,7 +101,7 @@ void AudioOutputWave::SetCurrentDevice(std::string deviceID){
 
 	bool wasPlaying=isPlaying;
 	isPlaying=false;
-	LOGV("closing, hWaveOut=%d", (int)hWaveOut);
+	LOGV("closing, hWaveOut=%lld", (long long)hWaveOut);
 	if(hWaveOut){
 		MMRESULT res;
 		if(isPlaying){
